Refactor(Helpers): image upload function and cloudinary function
