--- src/helpers/fileFilter.js
@@ -1,9 +1,0 @@
-/** @format */
-
-export const fileFilter = (req, file, cb) => {
-	if (file.mimetype.startsWith("image")) {
-		cb(null, true);
-	} else {
-		cb("invalid image file!", false);
-	}
-};
