+++ src/helpers/multer.js
@@ -1,6 +1,13 @@
 /** @format */
+import cloudinary from "./cloudinary.js";
 
-import cloudinary from "./imageUpload.js";
+export const fileFilter = (req, file, cb) => {
+	if (file.mimetype.startsWith("image")) {
+		cb(null, true);
+	} else {
+		cb("invalid image file!", false);
+	}
+};
 export const fileUpload = async (req) => {
 	let imageUrl = "";
 	await cloudinary.v2.uploader.upload(
